+++ tests/basic-proto/build.rs
@@ -379,15 +379,11 @@ fn static_lifetime_fields() {
 
 fn recursive() {
     let mut generator = Generator::new();
-    generator.configure(
-        ".Recursive.recursive",
-        Config::new().boxed(true).encoded_max_size(5),
+    generator.configure_many(
+        &[".Recursive.recursive", ".Recursive.rec"],
+        Config::new().recursive_field(),
     );
     generator.configure(".Recursive.of", Config::new().boxed(true));
-    generator.configure(
-        ".Recursive.rec",
-        Config::new().boxed(true).encoded_max_size(0),
-    );
     generator
         .compile_protos(
             &["proto/recursive.proto"],