+++ tests/basic-proto/src/recursive.rs
@@ -1,4 +1,3 @@
-use micropb::MessageEncode;
 use proto::Recursive;
 
 mod proto {
@@ -21,7 +20,4 @@ fn recursive_types() {
             let _: i32 = i;
         }
     }
-
-    // Encoded max size of first recursive field (set to 5) + max size of enum (size of int32 field)
-    assert_eq!(proto::Recursive::MAX_SIZE, Some(5 + (1 + 10)))
 }